Squashed commit of the following:
commit 11d0d3b8ff6bf308ef20d9dd598016c38ac4e33f
Author: Michael Blocksome <blocksom@us.ibm.com>
Date:   Wed Feb 3 16:35:43 2010 -0600

    Implement new queue interface for bgq, mpi, and lapiunix targets

commit cf375f8cbf4a771b19ad2896b007ff5bd30778b2
Author: Michael Blocksome <blocksom@us.ibm.com>
Date:   Wed Feb 3 16:00:56 2010 -0600

    Implement new queue interface for socklinux target

commit 065af118adc19b5b79278626d1cdd4e509449214
Author: Michael Blocksome <blocksom@us.ibm.com>
Date:   Wed Feb 3 15:37:45 2010 -0600

    Implement new queue interface for bgp target

commit 115a9f028fffee593ce3e24744b318b0b7f7e9aa
Author: Michael Blocksome <blocksom@us.ibm.com>
Date:   Wed Feb 3 11:16:56 2010 -0600

    Queue interface updates.

+++ test/multisend/multicast_global.cc
@@ -184,7 +184,7 @@ int main(int argc, char ** argv)
     mcast.src = (xmi_pipeworkqueue_t *)_buffer.srcPwq();
     mcast.dst = (xmi_pipeworkqueue_t *) _buffer.dstPwq();
 
-    mcast.client = client;	// client ID
+    mcast.client = (size_t) client;	// client ID
     mcast.context = 0;	// context ID
     mcast.roles = -1;
     mcast.bytes = TEST_BUF_SIZE;
